Create primitive classes in function
Each primitive is going to need its own class because I need a different token for each primitive to parse serialised programs. Create a function that defines a new class, separately specifying its token. This seems like a great solution except it can't be pickled, because now the class is local to the function...